requirements.txt
fastapi==0.111.0
uvicorn==0.18.2
openai==0.28.0
python-dotenv==0.21.0
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.3.0
pandas==2.2.2
openpyxl==3.1.2
tiktoken==0.5.2     
